Function clase::regex::agregar_pasos

pub fn agregar_pasos(
    steps: &mut Vec<PasoRegex>,
    chars_iter: &mut Chars<'_>
) -> Result<Vec<PasoRegex>, Error>

Agrega los pasos a la expresión regular.

Si el caracter es un punto, se agrega un paso con un comodín.

Si el caracter es un literal, se agrega un paso con el literal.

Si el caracter es una llave abierta, se obtiene el contenido de la llave y se agrega un paso con la cantidad de repeticiones y según si se trata de una repetición exacta, con solo mínimo, con solo máximo o ambas.

Si el caracter es un corchete abierto, se obtiene el contenido del corchete y se agrega un paso con la clase de caracteres.

Si el caracter es un asterisco, se agrega un paso con Alguna cantidad de repeticiones.

Si el caracter es un signo de interrogación, se agrega un paso con la cantidad de repeticiones (0 o 1 vez).

Si el caracter es un signo de más, se agrega un paso con la cantidad de repeticiones (1 o más).

Si el caracter es una barra, se obtiene el siguiente caracter y se agrega un paso con el literal.

Si el caracter es un dolar, se agrega un paso con un dolar.

Si el caracter es un caret, no se agrega un paso.

Si el caracter es una función OR, no se agrega un paso.

Si el caracter no es procesable, se devuelve un error.